rov_controller/requirements.txt
pyserial~=3.5.0
opencv-python~=4.11.0.86
